fixed refactor problems
