+++ JpaAdapter/src/main/java/be/xplore/notifyme/jpaadapters/JpaCommunicationPreferenceAdapter.java
@@ -1,12 +1,12 @@
 package be.xplore.notifyme.jpaadapters;
 
-import be.xplore.notifyme.services.communicationstrategies.ICommunicationStrategy;
 import be.xplore.notifyme.domain.CommunicationPreference;
 import be.xplore.notifyme.exceptions.JpaNotFoundException;
 import be.xplore.notifyme.jpaobjects.JpaCommunicationPreference;
 import be.xplore.notifyme.jparepositories.JpaCommunicationPreferenceRepository;
 import be.xplore.notifyme.jparepositories.JpaUserRepository;
 import be.xplore.notifyme.persistence.ICommunicationPreferenceRepo;
+import be.xplore.notifyme.services.communicationstrategies.ICommunicationStrategy;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
