+++ domain/src/test/java/be/xplore/notifyme/services/FacilityServiceTest.java
@@ -1,4 +1,4 @@
-package be.xplore.notifyme.communication;
+package be.xplore.notifyme.services;
 
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
