+++ rest-api/src/main/java/be/xplore/notifyme/controller/OrganisationManagerController.java
@@ -1,12 +1,12 @@
 package be.xplore.notifyme.controller;
 
-import be.xplore.notifyme.services.IUserOrgApplicationService;
-import be.xplore.notifyme.services.IUserOrgService;
-import be.xplore.notifyme.services.IUserService;
 import be.xplore.notifyme.dto.ApplicationUsernameDto;
 import be.xplore.notifyme.dto.OrganisationDto;
 import be.xplore.notifyme.dto.OrganisationLimitedInfoDto;
 import be.xplore.notifyme.dto.UserApplicationResponseDto;
+import be.xplore.notifyme.services.IUserOrgApplicationService;
+import be.xplore.notifyme.services.IUserOrgService;
+import be.xplore.notifyme.services.IUserService;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
