+++ rest-api/src/main/java/be/xplore/notifyme/controller/UserController.java
@@ -1,10 +1,5 @@
 package be.xplore.notifyme.controller;
 
-import be.xplore.notifyme.services.IKeycloakCommunicationService;
-import be.xplore.notifyme.services.IOrganisationService;
-import be.xplore.notifyme.services.IUserOrgApplicationService;
-import be.xplore.notifyme.services.IUserService;
-import be.xplore.notifyme.services.NotificationService;
 import be.xplore.notifyme.domain.CommunicationPreference;
 import be.xplore.notifyme.dto.ApplicationOrgNameDto;
 import be.xplore.notifyme.dto.NotificationDto;
@@ -14,6 +9,11 @@ import be.xplore.notifyme.dto.communicationpreference.GetCommunicationPreference
 import be.xplore.notifyme.dto.communicationpreference.PostCommunicationPreferenceDto;
 import be.xplore.notifyme.dto.communicationpreference.UpdateCommunicationPreferenceDto;
 import be.xplore.notifyme.services.ICommunicationPreferenceService;
+import be.xplore.notifyme.services.IKeycloakCommunicationService;
+import be.xplore.notifyme.services.IOrganisationService;
+import be.xplore.notifyme.services.IUserOrgApplicationService;
+import be.xplore.notifyme.services.IUserService;
+import be.xplore.notifyme.services.NotificationService;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
