+++ rest-api/src/test/java/be/xplore/notifyme/controller/AdminControllerTest.java
@@ -8,10 +8,6 @@ import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 
-import be.xplore.notifyme.communication.KeycloakCommunicationService;
-import be.xplore.notifyme.communication.OrganisationService;
-import be.xplore.notifyme.communication.UserService;
-import be.xplore.notifyme.communication.VenueService;
 import be.xplore.notifyme.config.KeycloakSecurityConfig;
 import be.xplore.notifyme.config.RestConfig;
 import be.xplore.notifyme.domain.Organisation;
@@ -19,6 +15,10 @@ import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.domain.Venue;
 import be.xplore.notifyme.exception.CrudException;
 import be.xplore.notifyme.exception.GeneralExceptionHandler;
+import be.xplore.notifyme.services.KeycloakCommunicationService;
+import be.xplore.notifyme.services.OrganisationService;
+import be.xplore.notifyme.services.UserService;
+import be.xplore.notifyme.services.VenueService;
 import java.util.HashSet;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
