+++ rest-api/src/test/java/be/xplore/notifyme/controller/OrganisationManagerControllerTest.java
@@ -9,9 +9,6 @@ import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 
-import be.xplore.notifyme.communication.UserOrgApplicationService;
-import be.xplore.notifyme.communication.UserOrgService;
-import be.xplore.notifyme.communication.UserService;
 import be.xplore.notifyme.config.KeycloakSecurityConfig;
 import be.xplore.notifyme.config.RestConfig;
 import be.xplore.notifyme.domain.OrgApplicationStatus;
@@ -20,6 +17,9 @@ import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.domain.UserOrgApplication;
 import be.xplore.notifyme.exception.GeneralExceptionHandler;
 import be.xplore.notifyme.exception.UnauthorizedException;
+import be.xplore.notifyme.services.UserOrgApplicationService;
+import be.xplore.notifyme.services.UserOrgService;
+import be.xplore.notifyme.services.UserService;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
