+++ rest-api/src/test/java/be/xplore/notifyme/controller/VenueManagerControllerTest.java
@@ -9,10 +9,6 @@ import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 
-import be.xplore.notifyme.communication.EventService;
-import be.xplore.notifyme.communication.FacilityService;
-import be.xplore.notifyme.communication.LineService;
-import be.xplore.notifyme.communication.VenueService;
 import be.xplore.notifyme.config.KeycloakSecurityConfig;
 import be.xplore.notifyme.config.RestConfig;
 import be.xplore.notifyme.domain.Event;
@@ -23,6 +19,10 @@ import be.xplore.notifyme.domain.Venue;
 import be.xplore.notifyme.dto.CreateEventDto;
 import be.xplore.notifyme.exception.CrudException;
 import be.xplore.notifyme.exception.GeneralExceptionHandler;
+import be.xplore.notifyme.services.EventService;
+import be.xplore.notifyme.services.FacilityService;
+import be.xplore.notifyme.services.LineService;
+import be.xplore.notifyme.services.VenueService;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashSet;
